requirements.txt
aiosqlite
authlib
dotenv
fastapi>=0.100.0
google-cloud-storage
itsdangerous
jinja2
openai
pillow
pytest
pytest-asyncio
pytest-cov
python-multipart
starlette>=0.27.0
uvicorn
